+++ common/src/main/java/mca/item/ExtendedWrittenBookItem.java
@@ -36,8 +36,9 @@ public class ExtendedWrittenBookItem extends WrittenBookItem {
 
     @Override
     public void appendTooltip(ItemStack stack, @Nullable World world, List<Text> tooltip, TooltipContext context) {
-        String key = String.format("mca.books.%s.author", book.getBookName());
-        tooltip.add(new TranslatableText(key).formatted(Formatting.GRAY));
+        if (book.getBookAuthor() != null) {
+            tooltip.add(book.getBookAuthor());
+        }
     }
 
     @Override