+++ changelog.md
@@ -189,3 +189,4 @@
 * Fixed issues when assigning family in editor
 * Buildings now support modded chests
 * Villagers will now use your editor name
+* Fixed letter author and creative mode usage